+++ server/listener/mdns_name_test.go
@@ -1,4 +1,4 @@
-package unifiedserver
+package listener
 
 import (
 	"testing"